requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
class-registry==2.1.2
Django==4.2.7
django-filter==23.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
filters==1.3.2
idna==3.4
inflection==0.5.1
packaging==23.2
PyJWT==2.8.0
pyotp==2.9.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.1.0
